requirements.txt
routeros-api==0.18.1
paho-mqtt==2.1.0
python-dotenv
